feat: canvgライブラリの追加と画像変換機能の整理
