+++ assets/template.json
@@ -9,7 +9,7 @@
                     "y": 10
                 },
                 "width": 190,
-                "height": 257,
+                "height": 277,
                 "rotate": 0,
                 "opacity": 1,
                 "borderWidth": 6,
